+++ sqlparser/lib/src/analysis/analysis.dart
@@ -3,6 +3,7 @@ import 'dart:math' show min, max;
 import 'package:meta/meta.dart';
 import 'package:source_span/source_span.dart';
 import 'package:sqlparser/sqlparser.dart';
+import 'package:sqlparser/src/engine/options.dart';
 import 'package:sqlparser/src/reader/tokenizer/token.dart';
 import 'package:sqlparser/src/utils/meta.dart';