+++ sqlparser/lib/src/analysis/context.dart
@@ -16,10 +16,11 @@ class AnalysisContext {
   /// A resolver that can be used to obtain the type of a [Typeable]. This
   /// mostly applies to [Expression]s, [Reference]s, [Variable]s and
   /// [ResultSet.resolvedColumns] of a select statement.
-  final TypeResolver types = TypeResolver();
+  final TypeResolver types;
 
   /// Constructs a new analysis context from the AST and the source sql.
-  AnalysisContext(this.root, this.sql);
+  AnalysisContext(this.root, this.sql, EngineOptions options)
+      : types = TypeResolver(options);
 
   /// Reports an analysis error.
   void reportError(AnalysisError error) {